updated files to merge more easily and removed unused code
