+++ include/sched.h
@@ -11,30 +11,12 @@
 
 #define TASK_RUNNING 0
 
+#include "event.h"
 #include "queue.h"
-#include "threads.h"
 
 extern struct task_struct* current;
 extern struct task_struct* task[NR_TASKS];
 extern int nr_tasks;
-
-namespace old {
-struct cpu_context {
-    unsigned long x19;
-    unsigned long x20;
-    unsigned long x21;
-    unsigned long x22;
-    unsigned long x23;
-    unsigned long x24;
-    unsigned long x25;
-    unsigned long x26;
-    unsigned long x27;
-    unsigned long x28;
-    unsigned long fp;
-    unsigned long sp;
-    unsigned long pc;
-};
-}  // namespace old
 
 struct task_struct {
     struct cpu_context cpu_context;
